# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://laravel-api.test",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a new product",
                "description": "Create a new product",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product by ID",
                "description": "Returns a single product",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing product",
                "description": "Updates a product and returns the updated product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "stock_quantity",
                                    "summary",
                                    "description",
                                    "category_id",
                                    "feature_image_url"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Product Title"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "stock_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Updated summary of the product."
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description of the product."
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "feature_image_url": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "http://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete an existing product",
                "description": "Deletes a product and returns a success message",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Product Delete Successful"
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreProductRequest": {
                "title": "StoreProductRequest",
                "description": "Request body for storing a new product",
                "properties": {
                    "title": {
                        "description": "The title of the product",
                        "type": "string",
                        "example": "Product Title"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "stock_quantity": {
                        "description": "The stock quantity of the product",
                        "type": "integer",
                        "example": 100
                    },
                    "summary": {
                        "description": "The summary of the product",
                        "type": "string",
                        "example": "Product summary"
                    },
                    "description": {
                        "description": "The description of the product",
                        "type": "string",
                        "example": "Product description"
                    },
                    "category_id": {
                        "description": "The ID of the category associated with the product",
                        "type": "integer",
                        "example": 1
                    },
                    "feature_image_url": {
                        "description": "The feature image URL of the product",
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the product",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "The title of the product",
                        "type": "string",
                        "example": "Product Title"
                    },
                    "slug": {
                        "description": "The slug of the product",
                        "type": "string",
                        "example": "product-title"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "stock_quantity": {
                        "description": "The stock quantity of the product",
                        "type": "integer",
                        "example": 100
                    },
                    "summary": {
                        "description": "The summary of the product",
                        "type": "string",
                        "example": "Product summary"
                    },
                    "feature_image_url": {
                        "description": "The feature image URL of the product",
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "status": {
                        "description": "The status of the product",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "The ID of the category associated with the product",
                        "type": "integer",
                        "example": 1
                    },
                    "category_type": {
                        "description": "The type of the category associated with the product",
                        "type": "string",
                        "example": "App\\Models\\Category"
                    },
                    "description": {
                        "description": "The description of the product",
                        "type": "string",
                        "example": "Product description"
                    },
                    "created_at": {
                        "description": "The date and time when the product was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-07 12:00:00"
                    },
                    "updated_at": {
                        "description": "The date and time when the product was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-07 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}